requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.1.1
pip==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
setuptools==69.2.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
wheel==0.43.0
WTForms==3.1.2